Optimize timers with exp == 0
Timer callbacks with an expiration of 0 are async calls via the event loop.
These don't actually need to call clock_gettime in most cases.

Warning: Changes the behavior for default created timers!
To use a timer with expiration time of 0 as a clock, start() needs to be called
to have a reference for elapsed().